refactor: 优化代码格式和类型安全
- 修复 CalAnomaly.py 中的类型检查问题，正确处理 Yanagi 类型转换
- 优化 Yanagi/__init__.py 中的 sim_instance None 检查
- 更新 pyproject.toml 中的 ruff 配置，移到 lint 部分
- 将所有 assert False 替换为 raise AssertionError 提高代码安全性
- 优化 Calculator.py 中的异常处理，保留异常链信息
- 使用 Python 3.10+ 的 strict=True 参数优化 zip 函数调用
- 清理过时的改进计划文档 (IMPROVEMENT_PLAN.md, IMPROVEMENT_SUMMARY.md)
- 重构事件处理器目录结构，移到独立的 event_handlers/ 子目录
