+++ pyproject.toml
@@ -53,6 +53,12 @@ dependencies = [
 [tool.ruff]
 line-length = 100
 
+[tool.ruff.lint]
+select = ["E", "F", "W", "B", "B9", "I", "Q"]
+ignore = ["E501"]
+exclude = ["build", "dist", ".venv", "__pypackages__"]
+per-file-ignores = { "__init__.py" = ["F401"] }
+
 [[tool.uv.index]]
 url = "https://pypi.tuna.tsinghua.edu.cn/simple"
 default = true
