+++ zsim/sim_progress/Character/Yanagi/__init__.py
@@ -2,6 +2,7 @@ from typing import TYPE_CHECKING
 
 from zsim.sim_progress.anomaly_bar.CopyAnomalyForOutput import NewAnomaly
 from zsim.sim_progress.Buff.BuffAddStrategy import buff_add_strategy
+from zsim.sim_progress.Preload import SkillNode
 
 from ..character import Character
 from ..utils.filters import (
@@ -9,8 +10,6 @@ from ..utils.filters import (
     _skill_node_filter,
 )
 from .StanceManager import StanceManager
-
-from zsim.sim_progress.Preload import SkillNode
 
 if TYPE_CHECKING:
     pass
@@ -32,14 +31,16 @@ class Yanagi(Character):
         for nodes in skill_nodes:
             self.stance_manager.update_myself(nodes)
         if self.cinema >= 1 and anomalies:
-            buff_add_strategy(self.cinme_1_buff_index, sim_instance=self.sim_instance)
+            if self.sim_instance is not None:
+                buff_add_strategy(self.cinme_1_buff_index, sim_instance=self.sim_instance)
             if self.cinema >= 4:
                 for _anomaly in anomalies:
                     if isinstance(_anomaly.activate_by, SkillNode):
                         if str(self.CID) in _anomaly.activate_by.skill_tag:
-                            buff_add_strategy(
-                                self.cinema_4_buff_index, sim_instance=self.sim_instance
-                            )
+                            if self.sim_instance is not None:
+                                buff_add_strategy(
+                                    self.cinema_4_buff_index, sim_instance=self.sim_instance
+                                )
                             break
 
     def update_sp_and_decibel(self, *args, **kwargs):
