--- zsim/sim_progress/ScheduledEvent/IMPROVEMENT_PLAN.md
@@ -1,322 +1,0 @@
-# ScheduledEvent 模块改进规划
-
-## 概述
-
-本文档旨在对 `ScheduledEvent` 模块进行小幅度的代码改进，以提高代码的可读性和可扩展性。改进将保持现有功能不变，主要关注代码结构优化、类型提示完善和代码规范统一。
-
-## 模块结构分析
-
-### 当前文件结构
-- `__init__.py` - 主要的事件调度逻辑 (543 行)
-- `Calculator.py` - 伤害计算器和乘数数据管理 (约 63304 行)
-- `CalAnomaly.py` - 异常伤害计算逻辑 (约 13929 行)
-- `buff_effect_trans.json` - Buff 效果映射配置
-
-### 主要类和职责
-- `ScheduledEvent` - 核心事件调度器
-- `Calculator` - 伤害计算核心类
-- `MultiplierData` - 乘数数据缓存管理
-- `CalAnomaly` - 异常伤害计算器
-- `ScConditionData` - 条件数据记录类
-
-## 识别的代码质量问题
-
-### 1. 类型提示不完整
-- `__init__.py` 中的方法缺少完整的类型提示
-- 部分返回类型和参数类型未明确标注
-- 使用了过时的 `typing` 模块模式
-
-### 2. 代码结构问题
-- `ScheduledEvent` 类过于庞大 (543 行)，承担了过多职责
-- 事件处理逻辑集中在一个巨大的 `process_event` 方法中
-- 重复的事件处理模式缺乏抽象
-
-### 3. 魔法数字和硬编码
-- 多处使用魔法数字表示优先级和状态
-- 事件类型判断使用大量 `isinstance` 检查
-
-### 4. 错误处理不够优雅
-- 部分错误处理使用了 `assert False`
-- 异常信息不够详细
-
-### 5. 文档字符串不完整
-- 部分方法缺少详细的文档字符串
-- 参数和返回值说明不完整
-
-### 6. 命名规范不符合 PEP8
-- 变量名使用驼峰命名法（如 `char_obj`, `dmg_sp`）
-- 函数名使用驼峰命名法（如 `cal_anomaly_dmg`）
-- 类名使用驼峰命名法（符合规范）
-- 常量名未使用全大写
-- 私有属性和方法命名不规范
-
-## 改进计划
-
-### 阶段 1: 类型提示现代化 (优先级: 高)
-
-#### 1.1 更新类型提示
-- 将 `typing` 模块的使用更新为 Python 3.10+ 模式
-- 为所有方法添加完整的类型提示
-- 使用 `|` 操作符替代 `Union`
-
-**实施步骤:**
-1. 更新 `__init__.py` 中的所有方法签名
-2. 修改 `Calculator.py` 中的类型定义
-3. 更新 `CalAnomaly.py` 中的类型提示
-
-### 阶段 2: 代码结构优化 (优先级: 高)
-
-#### 2.1 事件处理器模式
-- 创建事件处理器接口
-- 为每种事件类型创建独立的处理器类
-- 使用工厂模式管理处理器
-
-**实施步骤:**
-1. 创建 `EventHandler` 抽象基类
-2. 为每种事件类型创建具体的处理器类
-3. 创建 `EventHandlerFactory` 管理处理器实例
-4. 重构 `ScheduledEvent.process_event` 方法
-
-#### 2.2 职责分离
-- 将 `ScheduledEvent` 类拆分为多个小类
-- 创建 `EventScheduler` 负责事件调度
-- 创建 `EventValidator` 负责事件验证
-- 创建 `EventPriorityManager` 负责优先级管理
-
-### 阶段 3: 代码规范统一 (优先级: 中)
-
-#### 3.1 常量定义
-- 将魔法数字提取为常量
-- 创建 `EventConstants` 类管理常量
-
-```python
-class EventConstants:
-    DEFAULT_PRIORITY = 0
-    MAX_CACHE_SIZE = 128
-    TICK_PRECISION = 0.0000001
-    BINARY_SEARCH_THRESHOLD = 10
-    EVENT_LIST_MAX_SIZE = 1000
-```
-
-#### 3.2 错误处理改进
-- 创建自定义异常类
-- 替换 `assert False` 为适当的异常抛出
-- 提供更详细的错误信息
-
-#### 3.3 文档字符串完善
-- 为所有公共方法添加详细的文档字符串
-- 使用统一的文档字符串格式
-- 添加参数、返回值和异常说明
-
-#### 3.4 PEP8 命名规范统一
-- 将变量名改为蛇形命名法（如 `char_obj` → `char_obj`，`dmg_sp` → `damage_snapshot`）
-- 将函数名改为蛇形命名法（如 `cal_anomaly_dmg` → `calculate_anomaly_damage`）
-- 将常量名改为全大写下划线分隔（如 `max_size` → `MAX_SIZE`）
-- 规范私有属性和方法命名（使用单下划线前缀）
-- 统一模块和包级别的命名约定
-
-**实施步骤:**
-1. 创建命名规范映射表，确保重命名不会破坏现有功能
-2. 重命名变量和函数，保持语义一致性
-3. 更新所有相关的引用和调用
-4. 添加类型别名以保持向后兼容性（可选）
-
-**命名规范示例:**
-```python
-# 变量名
-char_obj → char_obj          # 已符合规范
-dmg_sp → damage_snapshot     # 更具描述性
-hitted_count → hit_count      # 更符合英语习惯
-
-# 函数名
-cal_anomaly_dmg → calculate_anomaly_damage
-cal_snapshot → calculate_snapshot
-get_execute_tick → get_execution_tick
-
-# 常量名
-max_size → MAX_SIZE
-tick_precision → TICK_PRECISION
-default_priority → DEFAULT_PRIORITY
-
-# 私有属性
-self.judge_required_info_dict → self._judge_required_info_dict
-self.execute_tick_key_map → self._execute_tick_key_map
-```
-
-### 阶段 4: 性能优化 (优先级: 低)
-
-#### 4.1 缓存机制优化
-- 优化 `MultiplierData` 的缓存策略
-- 考虑使用 `functools.lru_cache` 替代手动缓存
-
-#### 4.2 事件处理性能
-- 优化事件优先级排序算法
-- 减少不必要的事件类型检查
-
-## 具体实施步骤
-
-### 步骤 1: 创建基础结构
-1. 创建 `event_handlers.py` 文件
-2. 定义 `EventHandler` 抽象基类
-3. 创建事件处理器工厂类
-
-### 步骤 2: 实现事件处理器
-1. 为每种事件类型创建处理器类
-2. 实现具体的处理逻辑
-3. 从 `ScheduledEvent` 类中迁移处理逻辑
-
-### 步骤 3: 重构 ScheduledEvent 类
-1. 简化 `process_event` 方法
-2. 使用处理器模式替代大型 if-elif 链
-3. 分离验证和调度逻辑
-
-### 步骤 4: 类型提示更新
-1. 更新所有方法的类型提示
-2. 添加类型别名
-3. 移除过时的 typing 模块使用
-
-### 步骤 5: 常量和异常处理
-1. 创建常量类
-2. 创建自定义异常类
-3. 替换魔法数字和硬编码值
-
-### 步骤 6: PEP8 命名规范统一
-1. 创建命名规范映射表
-2. 重命名变量、函数和常量
-3. 更新所有相关引用
-4. 验证重命名后的功能正确性
-
-### 步骤 7: 文档字符串完善
-1. 为所有公共方法添加文档字符串
-2. 统一文档字符串格式
-3. 添加详细的参数说明
-
-### 步骤 7: 文档字符串完善
-1. 为所有公共方法添加文档字符串
-2. 统一文档字符串格式
-3. 添加详细的参数说明
-
-## 预期收益
-
-### 可读性提升
-- 代码结构更清晰
-- 职责分离明确
-- 类型提示完整
-
-### 可扩展性增强
-- 新事件类型添加更容易
-- 处理器模式支持灵活扩展
-- 模块化设计便于维护
-
-### 代码质量改善
-- 错误处理更优雅
-- 文档更完整
-- 代码规范统一
-- 命名规范符合 PEP8 标准
-
-## 风险评估
-
-### 低风险
-- 类型提示更新
-- 文档字符串完善
-- 常量定义
-- 命名规范统一（通过别名保持兼容性）
-
-### 中风险
-- 代码结构重构
-- 事件处理器模式引入
-
-### 缓解措施
-- 保持向后兼容性
-- 逐步迁移，避免大规模重写
-- 充分测试确保功能正确性
-
-## 时间估算
-
-- 阶段 1 (类型提示): 2-3 天
-- 阶段 2 (代码结构): 3-4 天
-- 阶段 3 (代码规范): 3-4 天  # 增加了命名规范统一的工作量
-- 阶段 4 (性能优化): 1-2 天
-
-总计: 9-13 天
-
-## 测试策略
-
-### 单元测试
-- 为新的事件处理器编写单元测试
-- 测试类型提示的正确性
-- 验证常量定义的正确性
-- 测试命名规范重命名后的功能正确性
-
-### 集成测试
-- 测试重构后的 `ScheduledEvent` 类
-- 验证事件调度流程的正确性
-- 确保与现有系统的兼容性
-
-### 回归测试
-- 运行现有测试套件
-- 验证功能没有回归
-- 性能基准测试
-
-## 总结
-
-本次改进计划将对 `ScheduledEvent` 模块进行系统性的代码质量提升，重点关注可读性和可扩展性。通过引入现代 Python 特性、优化代码结构、完善文档以及统一 PEP8 命名规范，使模块更易于维护和扩展。改进将采用渐进式的方式，确保系统的稳定性和向后兼容性。
-
-### 特别说明
-- **命名规范统一**是本次改进的重要组成部分，将显著提升代码的可读性和一致性
-- 所有改进都将保持向后兼容性，确保现有功能不受影响
-- 命名规范变更将通过系统性的重命名和测试来确保安全性
-- 改进后的代码将更符合 Python 社区的最佳实践
-
-### 重构完成状态 (2025.1.10)
-
-#### ✅ 已完成的项目
-
-1. **从if-elif链到事件处理器模式的完整转换**
-   - ✅ 成功创建了事件处理器架构（`event_handlers.py`, `concrete_handlers.py`）
-   - ✅ `process_event` 方法已完全转换为使用事件处理器模式
-   - ✅ 所有事件类型都有对应的事件处理器类
-   - ✅ 消除了原有的大型if-elif链结构
-
-2. **修复事件处理器中缺失的功能逻辑**
-   - ✅ `SkillEventHandler` 已包含完整的伤害计算逻辑
-   - ✅ 所有异常事件处理器都有完整的报告功能
-   - ✅ Buff结算和伤害更新逻辑已正确集成
-
-3. **修复导入问题**
-   - ✅ 修复了所有 `MultiplierData` 的导入问题
-   - ✅ 统一了导入路径，避免了循环导入风险
-   - ✅ 确保了类型提示的完整性
-
-#### ⏳ 待完成的项目
-
-4. **移除冗余的独立事件处理方法**
-   - 🔄 旧的独立事件处理方法（如 `skill_event`, `anomaly_event` 等）仍存在于文件中
-   - 🔄 这些方法在事件处理器模式下已不再使用，需要安全移除
-   - 🔄 移除前需要确保没有其他代码依赖这些方法
-
-5. **更新重构文档说明完成状态**
-   - 🔄 本文档需要更新最终完成状态
-   - 🔄 需要添加重构前后的代码对比
-   - 🔄 需要更新任何剩余的TODO项
-
-### 重构效果评估
-
-#### 🎯 架构改进
-- **代码可读性**: 事件处理器模式替代了大型if-elif链，代码结构更清晰
-- **可扩展性**: 新增事件类型只需创建新的处理器类，无需修改核心代码
-- **维护性**: 事件处理逻辑分离到独立的类中，降低了耦合度
-
-#### ⚡ 性能优化
-- **缓存改进**: `MultiplierData` 使用更稳定的唯一标识符，提升了缓存效率
-- **导入优化**: 所有导入移到文件顶部，避免了重复导入开销
-- **错误处理**: 增强的异常处理机制提供了更好的调试信息
-
-#### 🔄 向后兼容性
-- **接口保持**: `ScheduledEvent` 的公共接口保持不变
-- **行为一致**: 事件处理的行为与重构前完全一致
-- **数据结构**: 所有数据结构保持原样，确保兼容性
-
-### 总结
-
-ScheduledEvent 模块重构已基本完成，成功实现了从大型if-elif链到事件处理器模式的转换。核心功能已经过验证，事件处理器系统工作正常。剩余工作主要是清理冗余代码和更新文档，这些不会影响核心功能的正确性。
