--- zsim/sim_progress/ScheduledEvent/IMPROVEMENT_SUMMARY.md
@@ -1,168 +1,0 @@
-# ScheduledEvent 模块改进总结
-
-## 改进概述
-
-本次改进对 `ScheduledEvent` 模块进行了系统性的代码质量提升，重点关注可读性和可扩展性。通过引入现代 Python 特性、优化代码结构、完善文档以及统一 PEP8 命名规范，使模块更易于维护和扩展。同时修复了循环导入问题并确保类型安全。
-
-## 主要改进内容
-
-### 1. 事件处理器模式重构
-
-#### 1.1 创建事件处理器基础结构
-- **文件**: `event_handlers.py`
-- **改进**: 创建了 `EventHandler` 抽象基类和 `EventHandlerFactory` 工厂类
-- **好处**: 提供了统一的事件处理接口，支持灵活的处理器注册和管理
-
-#### 1.2 实现具体事件处理器类
-- **文件**: `concrete_handlers.py`
-- **改进**: 为每种事件类型创建了独立的处理器类：
-  - `SkillEventHandler` - 处理技能事件
-  - `AnomalyEventHandler` - 处理异常事件
-  - `DisorderEventHandler` - 处理紊乱事件
-  - `PolarityDisorderEventHandler` - 处理极性紊乱事件
-  - `AbloomEventHandler` - 处理薇薇安异放事件
-  - `RefreshEventHandler` - 处理数据刷新事件
-  - `QuickAssistEventHandler` - 处理快速支援事件
-  - `PreloadEventHandler` - 处理预加载事件
-  - `StunForcedTerminationEventHandler` - 处理眩晕强制终止事件
-  - `PolarizedAssaultEventHandler` - 处理极性强击事件
-- **好处**: 每个处理器职责单一，易于维护和扩展
-
-#### 1.3 重构 `process_event` 方法
-- **改进**: 将原有的大型 if-elif 链替换为事件处理器模式
-- **好处**: 
-  - 代码结构更清晰，可读性显著提升
-  - 新事件类型添加更容易
-  - 错误处理更加优雅
-
-### 2. 类型提示现代化
-
-#### 2.1 更新类型提示模式
-- **改进**: 将 `typing` 模块更新为 Python 3.10+ 模式
-  - 使用 `A | B` 替代 `Union[A, B]`
-  - 使用 `xxx | None` 替代 `Optional[xxx]`
-  - 使用 `list` 替代 `List[str]` 等旧式类型提示
-- **好处**: 代码更简洁，符合现代 Python 最佳实践
-
-#### 2.2 完善方法签名
-- **改进**: 为所有方法添加了完整的类型提示
-- **好处**: 提高了代码的类型安全性和 IDE 支持
-
-### 3. 常量定义统一
-
-#### 3.1 创建常量类
-- **文件**: `constants.py`
-- **改进**: 创建了 `EventConstants`、`ErrorMessages` 和 `WarningMessages` 类
-- **好处**: 集中管理常量，便于维护和修改
-
-### 4. PEP8 命名规范统一
-
-#### 4.1 私有属性规范化
-- **改进**: 将私有属性改为单下划线前缀
-  - `execute_tick_key_map` → `_execute_tick_key_map`
-- **好处**: 符合 Python 私有属性命名规范
-
-### 5. 文档字符串完善
-
-#### 5.1 统一文档字符串格式
-- **改进**: 所有公共方法都添加了详细的文档字符串
-- **格式**: 简介 + args + returns + raises
-- **好处**: 提高了代码的可读性和可维护性
-
-### 6. 缓存机制优化
-
-#### 6.1 优化 `MultiplierData` 缓存
-- **改进**: 
-  - 使用 `EventConstants.MAX_CACHE_SIZE` 统一缓存大小
-  - 引入 `@lru_cache` 装饰器优化缓存策略
-- **好处**: 提高了缓存管理的效率和可维护性
-
-## 架构改进
-
-### 改进前的架构问题
-- `ScheduledEvent` 类过于庞大（543 行），承担了过多职责
-- 事件处理逻辑集中在一个巨大的 `process_event` 方法中
-- 重复的事件处理模式缺乏抽象
-- 使用了大量的魔法数字和硬编码值
-
-### 改进后的架构优势
-- **职责分离**: 每个事件处理器只负责一种类型的事件
-- **可扩展性**: 新事件类型只需创建新的处理器类并注册即可
-- **可维护性**: 代码结构清晰，易于理解和修改
-- **可测试性**: 每个处理器都可以独立测试
-
-## 性能改进
-
-### 事件处理性能
-- 使用工厂模式管理处理器，减少运行时类型检查
-- 优化了事件优先级排序算法
-
-### 缓存机制
-- 统一了缓存大小管理
-- 引入了 LRU 缓存策略，提高缓存命中率
-
-## 向后兼容性
-
-所有改进都保持了向后兼容性，确保现有功能不受影响：
-- 命名规范变更通过映射管理
-- 事件处理器的引入保持了原有的处理逻辑
-- API 接口保持不变
-
-## 文件结构
-
-```
-zsim/sim_progress/ScheduledEvent/
-├── __init__.py                    # 主要的事件调度逻辑（已重构）
-├── Calculator.py                  # 伤害计算器（已优化）
-├── CalAnomaly.py                  # 异常伤害计算逻辑
-├── buff_effect_trans.json         # Buff 效果映射配置
-├── event_handlers.py              # 事件处理器基础结构（新增）
-├── concrete_handlers.py           # 具体事件处理器实现（新增）
-└── constants.py                   # 常量定义（新增）
-```
-
-## 测试建议
-
-### 单元测试
-- 为新的事件处理器编写单元测试
-- 测试类型提示的正确性
-- 验证常量定义的正确性
-
-### 集成测试
-- 测试重构后的 `ScheduledEvent` 类
-- 验证事件调度流程的正确性
-- 确保与现有系统的兼容性
-
-### 回归测试
-- 运行现有测试套件
-- 验证功能没有回归
-- 性能基准测试
-
-## 后续改进建议
-
-### 短期改进
-- 完善单元测试覆盖
-- 优化错误处理机制
-- 添加性能监控
-
-### 长期改进
-- 考虑引入异步事件处理
-- 实现事件处理的可视化调试
-- 添加事件处理的日志和监控
-
-## 总结
-
-本次改进成功地将 `ScheduledEvent` 模块从一个庞大的单体类重构为模块化、可扩展的事件处理系统。通过引入现代 Python 特性、优化代码结构、完善文档以及统一命名规范，显著提升了代码的可读性、可维护性和可扩展性。
-
-### 修复的重要问题
-1. **循环导入问题** - 通过将循环依赖的模块移到 `TYPE_CHECKING` 块中解决
-2. **类型安全** - 修复了所有类型错误，确保代码符合 Python 3.10+ 标准
-3. **未定义名称** - 在需要的地方添加了动态导入，解决 `Report` 和 `SkillNode` 引用问题
-
-### 验证结果
-- ✅ 所有模块导入成功
-- ✅ 常量定义正确
-- ✅ 事件处理器注册成功（共 10 个处理器）
-- ✅ 基本功能测试通过
-
-改进后的代码更符合 Python 社区的最佳实践，为后续的功能扩展和维护奠定了良好的基础。
