+++ zsim/sim_progress/ScheduledEvent/event_handlers/__init__.py
@@ -1,65 +1,25 @@
 """
 事件处理器模块
 
-该模块定义了事件处理的抽象基类和工厂类，用于管理各种类型的事件处理器。
+该模块定义了事件处理的工厂类，用于管理各种类型的事件处理器。
 """
 
-from abc import ABC, abstractmethod
 from typing import Any
 
-
-class EventHandler(ABC):
-    """事件处理器抽象基类"""
-
-    @abstractmethod
-    def can_handle(self, event: Any) -> bool:
-        """
-        判断是否可以处理指定类型的事件
-
-        Args:
-            event: 待处理的事件对象
-
-        Returns:
-            bool: 如果可以处理该类型事件则返回True，否则返回False
-        """
-        pass
-
-    @abstractmethod
-    def handle(self, event: Any, context: dict[str, Any]) -> None:
-        """
-        处理事件
-
-        Args:
-            event: 待处理的事件对象
-            context: 事件处理上下文，包含所需的数据和环境信息
-
-        Raises:
-            NotImplementedError: 如果子类未实现此方法
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def event_type(self) -> str:
-        """
-        返回处理器支持的事件类型名称
-
-        Returns:
-            str: 事件类型名称
-        """
-        pass
+from .base import EventHandlerABC
+from .concrete_handlers import register_all_handlers
 
 
 class EventHandlerFactory:
     """事件处理器工厂类"""
 
     def __init__(self):
-        self._handlers: dict[str, EventHandler] = {}
-        self._handler_cache: dict[type, EventHandler] = {}
+        self._handlers: dict[str, EventHandlerABC] = {}
+        self._handler_cache: dict[type, EventHandlerABC] = {}
         self._cache_hits = 0
         self._cache_misses = 0
 
-    def register_handler(self, handler: EventHandler) -> None:
+    def register_handler(self, handler: EventHandlerABC) -> None:
         """
         注册事件处理器
 
@@ -74,7 +34,7 @@ class EventHandlerFactory:
             raise ValueError(f"事件类型 '{event_type}' 的处理器已存在")
         self._handlers[event_type] = handler
 
-    def get_handler(self, event: Any) -> EventHandler | None:
+    def get_handler(self, event: Any) -> EventHandlerABC | None:
         """
         获取适合处理指定事件的处理器（带缓存）
 
@@ -100,7 +60,7 @@ class EventHandlerFactory:
         self._cache_misses += 1
         return None
 
-    def get_handler_by_type(self, event_type: str) -> EventHandler | None:
+    def get_handler_by_type(self, event_type: str) -> EventHandlerABC | None:
         """
         根据事件类型获取处理器
 
@@ -156,3 +116,5 @@ class EventHandlerFactory:
 
 # 全局处理器工厂实例
 event_handler_factory = EventHandlerFactory()
+
+__all__ = ["event_handler_factory", "register_all_handlers"]
